+++ src/main/java/com/address/book/addressbookapi/repo/MobileRepo.java
@@ -3,5 +3,5 @@ package com.address.book.addressbookapi.repo;
 import com.address.book.addressbookapi.entity.Mobile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MobileRepo extends JpaRepository<Mobile,Long> {
+public interface MobileRepo extends JpaRepository<Mobile,Integer> {
 }
