+++ src/main/java/com/address/book/addressbookapi/service/CustomerService.java
@@ -1,9 +1,6 @@
 package com.address.book.addressbookapi.service;
 
 import com.address.book.addressbookapi.dto.CustomerDto;
-import com.address.book.addressbookapi.entity.Customer;
-import com.address.book.addressbookapi.entity.ExternalContactEntity;
-import com.address.book.addressbookapi.entity.ExternalMobileEntity;
 
 import java.util.List;
 
@@ -16,5 +13,5 @@ public interface CustomerService {
 
     public CustomerDto saveDto(CustomerDto customerDto);
 
-    public String delete(Long id);
+    public String delete(Integer id);
 }
